LibGfx/JPEG2000: Move TagTree to its own file
No behavior change.
