+++ Meta/gn/secondary/Userland/Libraries/LibGfx/BUILD.gn
@@ -81,6 +81,7 @@ shared_library("LibGfx") {
     "ImageFormats/JBIG2Loader.cpp",
     "ImageFormats/JPEG2000Loader.cpp",
     "ImageFormats/JPEG2000ProgressionIterators.cpp",
+    "ImageFormats/JPEG2000TagTree.cpp",
     "ImageFormats/JPEGLoader.cpp",
     "ImageFormats/JPEGWriter.cpp",
     "ImageFormats/JPEGXLEntropyDecoder.cpp",
