+++ Userland/Libraries/LibGfx/ImageFormats/JPEG2000Loader.h
@@ -19,22 +19,6 @@ enum class SubBand {
     HorizontalHighpassVerticalHighpass, // "HH" in spec
 };
 
-struct TagTreeNode;
-class TagTree {
-public:
-    TagTree(TagTree&&);
-    ~TagTree();
-
-    static ErrorOr<TagTree> create(u32 x_count, u32 y_count);
-
-    ErrorOr<u32> read_value(u32 x, u32 y, Function<ErrorOr<bool>()> const& read_bit, Optional<u32> stop_at = {}) const;
-
-private:
-    TagTree(NonnullOwnPtr<TagTreeNode>);
-
-    NonnullOwnPtr<TagTreeNode> m_root;
-};
-
 }
 
 struct JPEG2000LoadingContext;
